Using cookies

+++ messages.go
@@ -9,7 +9,7 @@ type authorizationResp struct {
 	Error            string `json:"error"`
 	ErrorDescription string `json:"error_description"`
 	AccessToken      string `json:"access_token"`
-	ExpiresIn        int64  `json:"expires_in"`
+	ExpiresIn        int    `json:"expires_in"`
 	RefreshToken     string `json:"refresh_token"`
 	Scope            string `json:"scope"`
 	TokenType        string `json:"token_type"`
